Optimistically update the UI (#1)
* Optimistically update the UI. When user clicks on a card, if they're not
near to a server (and without optimistic UI updating), there's latency
between the click and the card being highlighted. This adds optimistic
updates so that the card is highlighted client-side and then the server
ensures the move was legal and sends the _actual_ game state back down
to the client.